% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the U.S. Federal Holidays Explorer Shiny Application}
\usage{
launch_app()
}
\value{
No return value, called for side effects
}
\description{
This function launches the Shiny application for exploring U.S. Federal Holidays.
It includes visualizations and information about current and proposed federal holidays.
}
\examples{
if (interactive()) {
  launch_app()
}
}
